Fix race in code
Assuming a process p1 already has a lock, if a 2nd process p2 is
trying to grab a lock and it successfully reaches the ownership
check...that is to say it failed to create a link to it's own pid
file, and it verified that the existing lockfile exists, between the
call to os.SameFile(fiTmp, fiLock) and the read in GetOwner the
lockfile can be deleted by p1.

This fix makes GetOwner returns a retryable error in the case where
the lockfile disappears before it can be verified for pid.